Add shorts-weather endpoint to API
